break games into componets (organize later)
