+++ src/components/games/ListGames.vue
@@ -1,69 +1,6 @@
 <script setup lang="ts">
-import { reactive, onMounted, computed } from "vue";
+import {items, select} from "./Games"
 
-interface Game {
-  name: string;
-  file: string;
-  root: string;
-  img: string;
-}
-
-const items = reactive({
-  all: [] as Game[],
-  page: [] as Game[],
-  pagenum: 1,
-  itemsperpage: 25,
-
-});
-
-// Initialize URL params in onMounted
-onMounted(() => {
-  fetchStuff();
-});
-
-async function fetchStuff() {
-  try {
-    const response = await fetch("/cdn/games.json");
-    const data = await response.json();
-    items.all = data.sort((a: { name: string }, b: { name: string }) =>
-      a.name.localeCompare(b.name)
-    );
-    items.search = items.urlParams.search;
-
-    page(items.urlParams.page);
-  } catch (error) {
-    console.error('Failed to fetch games:', error);
-  }
-}
-
-const totalPages = computed(() => {
-  if (items.urlParams.search)
-    return Math.ceil(lastSearchResultsLength.value / items.itemsperpage);
-  return Math.ceil(items.all.length / items.itemsperpage);
-});
-
-const lastSearchResultsLength = computed(() => {
-  return items.all.filter((game) =>
-    game.name.toLowerCase().includes(items.urlParams.search.toLowerCase())
-  ).length;
-});
-
-const filteredResults = computed(() => {
-  if (!items.search) return [];
-  return items.all.filter((game) =>
-    game.name.toLowerCase().includes(items.search.toLowerCase())
-  );
-});
-
-
-function select(file: string, root: string) {
-  const path = "/cdn/" + root + "/" + file;
-  if (root === "webretro") {
-    window.location.href = "/Iframe/cdn/webretro/index.html?core=autodetect&rom=" + file;
-  } else {
-    window.location.href = "/Iframe" + path;
-  }
-}
 
 </script>
