+++ src/pages/games.astro
@@ -6,7 +6,12 @@ import GamesPaging from "../components/games/GamesPaging.vue";
 ---
 
 <Navbar>
-  <!-- <SearchGames client:load />
+  <SearchGames client:load />
   <ListGames client:load />
-  <GamesPaging client:visible /> -->
+  <GamesPaging client:visible />
 </Navbar>
+
+<script>
+  import {mounted} from "../components/games/Games"
+  mounted()
+</script>
